Remove unused utility functions and test files to streamline the codebase
- Deleted the `slugify` function from `utils.ts` as it was no longer needed.
- Removed all test files related to articles, users, and index tests to clean up the testing suite.
- This cleanup helps in maintaining a more focused and efficient codebase.
