+++ src/shared/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./slugify.util";
+export * from "./tests.utils";
