+++ src/tags/tags.test.ts
@@ -3,26 +3,13 @@ import { treaty } from "@elysiajs/eden";
 import { app } from "@/core/app";
 import { db } from "@/core/db";
 
-function expectToBeDefined<T>(value: T | null | undefined): asserts value is T {
-	expect(value).toBeDefined();
-}
-
-function expectSuccess(response: { error: unknown; data: unknown }) {
-	expect(response.error).toBeNull();
-	expect(response.data).toBeDefined();
-}
-
+// Create type-safe API client with Eden Treaty
 const { api } = treaty(app);
 
+// Test data
 const testUser = {
 	email: "test@test.com",
 	username: "testuser",
-	password: "Password123",
-};
-
-const testUser2 = {
-	email: "celeb@test.com",
-	username: "celeb_testuser",
 	password: "Password123",
 };
 
@@ -34,33 +21,45 @@ const testArticle = {
 };
 
 let authToken: string;
-let authToken2: string;
-let articleSlug: string;
 
 beforeAll(async () => {
+	// Reset database
 	await db.$executeRaw`TRUNCATE TABLE users, articles, tags, comments CASCADE`;
 
+	// Register user
 	const reg1 = await api.users.post({ user: testUser });
 	authToken = reg1.data?.user?.token ?? "";
 
-	const reg2 = await api.users.post({ user: testUser2 });
-	authToken2 = reg2.data?.user?.token ?? "";
-
+	// Login user (to ensure token is valid)
 	const login1 = await api.users.login.post({
 		user: { email: testUser.email, password: testUser.password },
 	});
 	authToken = login1.data?.user?.token ?? "";
 
-	const login2 = await api.users.login.post({
-		user: { email: testUser2.email, password: testUser2.password },
+	// Create an article with tags to ensure tags exist
+	await api.articles.post(
+		{
+			article: testArticle,
+		},
+		{
+			headers: {
+				Authorization: `Token ${authToken}`,
+			},
+		},
+	);
+});
+
+describe("Tags Tests", () => {
+	afterAll(async () => {
+		await db.$disconnect();
 	});
-	authToken2 = login2.data?.user?.token ?? "";
-});
 
-afterAll(async () => {
-	await db.$disconnect();
-});
+	it("should get all tags", async () => {
+		const { data, error } = await api.tags.get();
 
-describe("Articles", () => {
-	// ... (PASTE ALL ARTICLE-RELATED TESTS FROM index.test.ts HERE)
+		expect(error).toBeNull();
+		expect(data?.tags).toBeDefined();
+		expect(Array.isArray(data?.tags)).toBe(true);
+		expect(data?.tags.length).toBeGreaterThan(0);
+	});
 });
